Implement immutable tree lists
Add various classes to implement immutable tree lists. The
AbstractImmutableTreeNode class can be subclassed to add additional
information to each node. The user must subclass
AbstractImmutableTreeList to specify the type of
AbstractImmutableTreeNode as a type parameter. The user must also
subclass AbstractImmutableTreeListFactory and
AbstractImmutableTreeNodeFactory to be able to create trees using his or
her subclasses.